move all plot components into a folder
