+++ src/components/plotStyling/AxesDesign.jsx
