+++ src/components/plotStyling/ColorBrowser.jsx
