+++ src/components/plotStyling/ColorPickerOption.jsx
