+++ src/components/plotStyling/DimensionsRangeEditor.jsx
