+++ src/components/plotStyling/LabelsDesign.jsx
