+++ src/components/plotStyling/LegendEditor.jsx
